refactor(#3037): relocate escalation-settings read path to services (backflow 15→14)
Move the escalation-settings resolution helpers out of the server route layer
into a new `crate::services::escalation_settings` module so service-layer
callers no longer reach back up via `crate::server::routes::escalation::*`.

Cluster (DTO/helper down-shift, lowest-risk coherent unit):
- `effective_owner_user_id_with_backends`, `merged_settings_pg`,
  `load_override_pg_async`, `escalation_defaults`, `normalize_optional_string`,
  and `ESCALATION_SETTINGS_OVERRIDE_KEY` move from
  `src/server/routes/escalation.rs` to `src/services/escalation_settings.rs`.
- These helpers depend only on `crate::config` types/constants (already below
  services) and `crate::utils::async_bridge` — no AppState, axum, or server
  types — so the relocation has no cyclic dependency.
- `src/services/dispatches/discord_delivery/orchestration.rs:42` now calls the
  service path; the escalation route re-imports the helpers as the single
  source of truth (no unused imports; dropped now-unused config imports).

Behavior unchanged: pure code relocation, identical config-defaults +
kv_meta-override resolution. Owner resolution feeds dispatch delivery and is
not leader-gated (same kv_meta read on leader and standby), so no new
multinode ownership/lease/singleton assumption — no Audited-touches entry
warranted.

Gates: cargo fmt --check clean; cargo check --lib / --lib --tests clean (0 new
warnings); escalation + discord_delivery lib tests pass (3 + 20); backflow
baseline 15→14; inventory-docs freshness passed; ci-script-checks exit 0.

Refs #3037
