+++ src/services/mod.rs
@@ -29,6 +29,7 @@ pub mod dispatched_sessions;
 pub mod dispatches;
 pub mod dispatches_followup;
 pub mod envelope_dedup;
+pub mod escalation_settings;
 pub mod gemini;
 pub mod git;
 pub mod issue_announcements;
